requirements.txt
Django>=4.2
psycopg2-binary>=2.9
watchdog
djangorestframework
django-cors-headers
google-genai

